add name and time_per_step methods
